adds files from 2023 dactyl attempt
added an oled and trying to use a pro micro instead of a teensy with no
success. I think the wiring is all correct, but all I have to check is a
multi meter. tried swapping col2row to row2col and it got worse. Maybe
the problem is somewhere in adereth's old matrix.c or maybe the io
expander is poop. need a way to test.